+++ db/User.js
@@ -1,6 +1,5 @@
 const mongoose = require("mongoose");
 const { Schema, model } = mongoose;
-const Joi = require("joi");
 
 const { handleSaveError, runValidatorsAtUpdate } = require("./hooks");
 const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
@@ -42,17 +41,8 @@ userSchema.pre("findOneAndUpdate", runValidatorsAtUpdate);
 
 userSchema.post("findOneAndUpdate", handleSaveError);
 
-const userSignupSchema = Joi.object({
-  username: Joi.string().required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const userSigninSchema = Joi.object({
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
 const User = model("user", userSchema); //создал автоматически коллекцию user на монго
 
-module.exports = { User, userSignupSchema, userSigninSchema };
+module.exports = {
+  User,
+};
