+++ db/contacts-schema.js
@@ -1,7 +1,5 @@
 const mongoose = require("mongoose");
 const { Schema, model } = mongoose;
-const Joi = require("joi");
-const myCustomJoi = Joi.extend(require("joi-phone-number"));
 const { handleSaveError, runValidatorsAtUpdate } = require("./hooks");
 
 const contactSchema = new Schema(
@@ -39,18 +37,8 @@ contactSchema.pre("findOneAndUpdate", runValidatorsAtUpdate);
 
 contactSchema.post("findOneAndUpdate", handleSaveError);
 
-const schema = Joi.object({
-  id: Joi.string(),
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: myCustomJoi.string().phoneNumber().required(),
-  favorite: Joi.boolean(),
-});
-
-const schemaFav = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
 const Contact = model("contact", contactSchema);
 
-module.exports = { Contact, schema, schemaFav };
+module.exports = {
+  Contact,
+};
